+++ nes_cpp/KbInputHandler.cpp
@@ -37,7 +37,7 @@ this program.  If not, see <http://www.gnu.org/licenses/>.
      void KbInputHandler::keyPressed(KeyEvent* ke) {
 
         int kc = ke->getKeyCode();
-        if (kc >= allKeysState->size()) {
+        if (kc >= (int)allKeysState->size()) {
             return;
         }
 
@@ -58,7 +58,7 @@ this program.  If not, see <http://www.gnu.org/licenses/>.
      void KbInputHandler::keyReleased(KeyEvent* ke) {
 
         int kc = ke->getKeyCode();
-        if (kc >= allKeysState->size()) {
+        if (kc >= (int)allKeysState->size()) {
             return;
         }
